Fix TODOs about length being UInt vs Int; other little warnings

+++ Applications/MIDIMonitor/DetailsWindowController.swift
@@ -76,7 +76,7 @@ class DetailsWindowController: UtilityWindowController, NSWindowDelegate {
 
     private func updateDescriptionFields() {
         let format = NSLocalizedString("%@ bytes", tableName: "MIDIMonitor", bundle: SMBundleForObject(self), comment: "Details size format string")
-        let formattedLength = SMMessage.formatLength(UInt(dataForDisplay.count))!
+        let formattedLength = SMMessage.formatLength(dataForDisplay.count)!
         let sizeString = String.localizedStringWithFormat(format, formattedLength)
 
         sizeField.stringValue = sizeString
